AxisSets

AxisSets.DEFAULT_FLATTEN_DELIM

AxisSets.DEFAULT_PROD_DELIM

AxisSets.KeyedDataset

AxisSets.KeyedDataset

AxisSets.Pattern

AxisKeys.axiskeys

AxisSets.constraintmap

AxisSets.dimpaths

AxisSets.flatten

AxisSets.validate

Base.getindex

Base.getproperty

Base.setindex!

NamedDims.dimnames

AxisSets.DEFAULT_FLATTEN_DELIM — Constant

DEFAULT_FLATTEN_DELIM

:⁻ (or \^-)

Separates the parent symbols of a nested NamedTuple that has been flattened. A less common symbol was used to avoid collisions with :_ in the parent symbols.

Example

julia> using AxisSets: flatten

julia> data = (
           val1 = (a1 = 1, a2 = 2),
           val2 = (b1 = 11, b2 = 22),
           val3 = [111, 222],
           val4 = 4.3,
       );

julia> flatten(data)
(val1⁻a1 = 1, val1⁻a2 = 2, val2⁻b1 = 11, val2⁻b2 = 22, val3 = [111, 222], val4 = 4.3)

AxisSets.DEFAULT_PROD_DELIM — Constant

DEFAULT_PROD_DELIM

:ᵡ (or \^x)

Separates the parent symbols from an n-dimensional array that was flattened / reshaped. A less common symbol was used to avoid collisions with :_ in the parent symbols.

Example

julia> using AxisKeys, Dates; using AxisSets: flatten

julia> A = KeyedArray(
           reshape(1:24, (4, 3, 2));
           time=DateTime(2021, 1, 1, 11):Hour(1):DateTime(2021, 1, 1, 14),
           obj=[:a, :b, :c],
           loc=[1, 2],
       );

julia> axiskeys(flatten(A, (:obj, :loc)), :objᵡloc)
6-element Array{Symbol,1}:
 :aᵡ1
 :bᵡ1
 :cᵡ1
 :aᵡ2
 :bᵡ2
 :cᵡ2

AxisSets.KeyedDataset — Type

KeyedDataset{T}

A KeyedDataset describes an associative collection of component KeyedArrays with constraints on their shared dimensions.

Fields

constraints::OrderedSet{Pattern} - Constraint Patterns on shared dimensions.

data::LittleDict{Tuple{Vararg{Symbol}}, T} - Flattened key paths as tuples of symbols to each component array of type T.

AxisSets.KeyedDataset — Method

(ds::KeyedDataset)(key) -> KeyedDataset

A collable syntax for selecting of filtering a subset of a KeyedDataset.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset, flatten;

julia> ds = KeyedDataset(
           flatten([
               :g1 => [
                   :a => KeyedArray(zeros(3); time=1:3),
                   :b => KeyedArray(ones(3, 2); time=1:3, loc=[:x, :y]),
                ],
                :g2 => [
                    :a => KeyedArray(ones(3); time=1:3),
                    :b => KeyedArray(zeros(3, 2); time=1:3, loc=[:x, :y]),
                ]
            ])...
       );

julia> collect(keys(ds(:__, :a).data))
2-element Array{Tuple{Vararg{Symbol,N} where N},1}:
 (:g1, :a)
 (:g2, :a)

julia> collect(keys(ds(:g1, :__).data))
2-element Array{Tuple{Vararg{Symbol,N} where N},1}:
 (:g1, :a)
 (:g1, :b)

AxisSets.Pattern — Type

Pattern

A pattern is just a wrapper around a Tuple{Vararg{Symbol}} which enables searching and filtering for matching components and dimension paths in a KeyedDataset. Special symbols :_ and :__ are used as wildcards, similar to * and ** in glob pattern matching.

Example

julia> using AxisSets: Pattern;

julia> items = [
           (:train, :input, :load, :time),
           (:train, :input, :load, :id),
           (:train, :input, :temperature, :time),
           (:train, :input, :temperature, :id),
           (:train, :output, :load, :time),
           (:train, :output, :load, :id),
       ];

julia> filter(in(Pattern(:__, :time)), items)
3-element Array{NTuple{4,Symbol},1}:
 (:train, :input, :load, :time)
 (:train, :input, :temperature, :time)
 (:train, :output, :load, :time)

julia> filter(in(Pattern(:__, :load, :_)), items)
4-element Array{NTuple{4,Symbol},1}:
 (:train, :input, :load, :time)
 (:train, :input, :load, :id)
 (:train, :output, :load, :time)
 (:train, :output, :load, :id)

AxisKeys.axiskeys — Method

axiskeys(ds)
axiskeys(ds, dimname)
axiskeys(ds, pattern)
axiskeys(ds, dimpath)

Returns a list of unique axis keys within the KeyedDataset. A Tuple will always be returned unless you explicitly specify the dimpath you want.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset;

julia> ds = KeyedDataset(
           :val1 => KeyedArray(rand(4, 3, 2); time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
           :val2 => KeyedArray(rand(4, 3, 2) .+ 1.0; time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
       );

julia> axiskeys(ds)
(1:4, -1:-1:-3, [:a, :b])

julia> axiskeys(ds, :time)
(1:4,)

julia> axiskeys(ds, (:val1, :time))
1:4

AxisSets.constraintmap — Method

constraintmap(ds)

Returns a mapping of constraint patterns to specific dimension paths. The returned dictionary has keys of type Pattern and the values are sets of Tuple{Vararg{Symbol}}.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset, constraintmap;

julia> ds = KeyedDataset(
           :val1 => KeyedArray(rand(4, 3, 2); time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
           :val2 => KeyedArray(rand(4, 3, 2) .+ 1.0; time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
       );

julia> collect(constraintmap(ds))
3-element Array{Pair{AxisSets.Pattern,Set{Tuple{Vararg{Symbol,N} where N}}},1}:
 AxisSets.Pattern((:__, :time)) => Set([(:val2, :time), (:val1, :time)])
  AxisSets.Pattern((:__, :loc)) => Set([(:val1, :loc), (:val2, :loc)])
  AxisSets.Pattern((:__, :obj)) => Set([(:val2, :obj), (:val1, :obj)])

AxisSets.dimpaths — Method

dimpaths(ds, [pattern]) -> Vector{<:Tuple{Vararg{Symbol}}}

Return a list of all dimension paths in the KeyedDataset. Optionally, you can filter the results using a Pattern.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset, dimpaths;

julia> ds = KeyedDataset(
           :val1 => KeyedArray(rand(4, 3, 2); time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
           :val2 => KeyedArray(rand(4, 3, 2) .+ 1.0; time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
       );

julia> dimpaths(ds)
6-element Array{Tuple{Symbol,Symbol},1}:
 (:val1, :time)
 (:val1, :loc)
 (:val1, :obj)
 (:val2, :time)
 (:val2, :loc)
 (:val2, :obj)

AxisSets.flatten — Function

flatten(collection, [delim])

Flatten a collection of nested associative types into a flat collection of pairs. If the input keys are symbols (ie: NamedTuple) then the ⁻ will be used, otherwise Tuple keys will be returned.

Example

julia> using AxisSets: flatten

julia> data = (
           val1 = (a1 = 1, a2 = 2),
           val2 = (b1 = 11, b2 = 22),
           val3 = [111, 222],
           val4 = 4.3,
       );

julia> flatten(data)
(val1⁻a1 = 1, val1⁻a2 = 2, val2⁻b1 = 11, val2⁻b2 = 22, val3 = [111, 222], val4 = 4.3)

flatten(A, dims, [delim])

Flatten a KeyedArray along the specified consecutive dimensions. The dims argument can either be a Tuple of symbols or a Pair{Tuple, Symbol} if you'd like to specify the desired flattened dimension name. If the dims is just a Tuple with no output dimension specified then ᵡ will be used to generate the new dimension name.

Example

julia> using AxisKeys, Dates, NamedDims; using AxisSets: flatten

julia> A = KeyedArray(
           reshape(1:24, (4, 3, 2));
           time=DateTime(2021, 1, 1, 11):Hour(1):DateTime(2021, 1, 1, 14),
           obj=[:a, :b, :c],
           loc=[1, 2],
       );

julia> dimnames(flatten(A, (:obj, :loc)))
(:time, :objᵡloc)

AxisSets.validate — Method

validate(ds, [constraint])

Validate that all constrained dimension paths within a KeyedDataset have matching key values. Optionally, you can test an explicit constraint Pattern.

Returns

true if an error isn't thrown

Throws

ArgumentError: If the constraints are not respected

Base.getindex — Method

getindex(ds::KeyedDataset, key)

Lookup KeyedDataset component by its Tuple key, or Symbol for keys of depth 1. Shared axis keys for the returned KeyedArray are wrapped in a ReadOnlyArray for safety.

Example

```jldoctest julia> using AxisKeys; using AxisSets: KeyedDataset;

julia> ds = KeyedDataset( :val1 => KeyedArray(zeros(3, 2); time=1:3, obj=[:a, :b]), :val2 => KeyedArray(ones(3, 2) .+ 1.0; time=1:3, loc=[:x, :y]), );

julia> ds[:val1] 2-dimensional KeyedArray(NamedDimsArray(...)) with keys: ↓ time ∈ 3-element ReadOnlyArrays.ReadOnlyArray{Int64,...} → obj ∈ 2-element ReadOnlyArrays.ReadOnlyArray{Symbol,...} And data, 3×2 Array{Float64,2}: (:a) (:b) (1) 0.0 0.0 (2) 0.0 0.0 (3) 0.0 0.0

Base.getproperty — Method

getproperty(ds::KeyedDataset, sym::Symbol)

Extract KeyedDataset fields, dimension keys or components in that order. Shared axis keys are wrapped in a ReadOnlyArray for safety.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset;

julia> ds = KeyedDataset(
           :val1 => KeyedArray(zeros(3, 2); time=1:3, obj=[:a, :b]),
           :val2 => KeyedArray(ones(3, 2) .+ 1.0; time=1:3, loc=[:x, :y]),
       );

julia> collect(keys(ds.data))
2-element Array{Tuple{Vararg{Symbol,N} where N},1}:
 (:val1,)
 (:val2,)

julia> ds.time
3-element ReadOnlyArrays.ReadOnlyArray{Int64,1,UnitRange{Int64}}:
 1
 2
 3

julia> dimnames(ds.val1)
(:time, :obj)

Base.setindex! — Method

setindex!(ds::KeyedDataset{T}, val, key) -> T

Store the new val in the KeyedDataset. If any new dimension names don't any existing constraints then Pattern(:__, <dimname>) is used by default. If the axis values of the new val doesn't meet the existing constraints in the dataset then an error will be throw.

Example

julia> using AxisKeys; using AxisSets: KeyedDataset, constraintmap;

julia> ds = KeyedDataset(:a => KeyedArray(zeros(3); time=1:3));

julia> ds[:b] = KeyedArray(ones(3, 2); time=1:3, lag=[-1, -2]);

julia> collect(constraintmap(ds))
2-element Array{Pair{AxisSets.Pattern,Set{Tuple{Vararg{Symbol,N} where N}}},1}:
 AxisSets.Pattern((:__, :time)) => Set([(:b, :time), (:a, :time)])
  AxisSets.Pattern((:__, :lag)) => Set([(:b, :lag)])

julia> ds[:c] = KeyedArray(ones(3, 2); time=2:4, lag=[-1, -2])
ERROR: ArgumentError: Shared dimensions don't have matching keys

NamedDims.dimnames — Method

dimnames(ds)

Returns a list of the unique dimension names within the KeyedDataset.

Example

julia> using AxisKeys; using NamedDims; using AxisSets: KeyedDataset;

julia> ds = KeyedDataset(
           :val1 => KeyedArray(rand(4, 3, 2); time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
           :val2 => KeyedArray(rand(4, 3, 2) .+ 1.0; time=1:4, loc=-1:-1:-3, obj=[:a, :b]),
       );

julia> dimnames(ds)
3-element Array{Symbol,1}:
 :time
 :loc
 :obj