AxisSets

AxisSets.Pattern

AxisSets.Pattern — Type

Patter

A pattern is just a wrapper around a Tuple{Vararg{Symbol}} which enables searching for matching components and dimension paths in a Dataset. Special symbols :_ and :__ are used as wildcards, similar to * and ** in glob pattern matching.